backend/requirements.txt
asgiref==3.8.1
certifi==2025.4.26
charset-normalizer==3.4.2
Deprecated==1.2.18
distro==1.9.0
Django==5.2
django-cors-headers==4.7.0
django-js-asset==3.1.2
django-mptt==0.17.0
djangorestframework==3.16.0
gunicorn==23.0.0
idna==3.10
netaddr==1.3.0
packaging==25.0
pillow==11.2.1
psycopg2-binary==2.9.10
python-dotenv==1.1.0
requests==2.32.4
sqlparse==0.5.3
urllib3==2.4.0
wrapt==1.17.2
yookassa==3.5.0
